Min.Me/006/Dependencies_Update_003
- Updated node modules, dependencies and yarn packages
- Added CI/CD scripts into project
