+++ .travis.yml
@@ -1,0 +1,7 @@
+language: node_js
+
+node_js:
+
+- node   
+
+- lts/*
