+++ docs/ISSUES_TEMPLATES.md
@@ -1,45 +1,38 @@
-# List of changes made: 
+# List of changes made:
 
 -
 -
--
-
--------------------------------------------------------------
+- 
 
 ## Type of Issue:
 
-* New feature request
-* Bug report-Major
-* Bug report-Medium
-* Bug report-Small
+- New feature request
+- Bug report-Major
+- Bug report-Medium
+- Bug report-Small
 
--------------------------------------------------------------
+---
+
 ## Recent issue description:
 
+---
 
-
--------------------------------------------------------------
 ### Possible solution:
 
+---
 
-
--------------------------------------------------------------
 ### Step by step guide for recent PR:
 
 1.
-2. 
+2.
 3.
 
+---
 
--------------------------------------------------------------
 ### Context of current issue:
 
+---
 
-
--------------------------------------------------------------
 ### Your current work environment:
 
-
-
-
--------------------------------------------------------------
+---
