+++ src/components/Settings.js
@@ -1,10 +1,9 @@
-import '../App.css';
+import "../App.css";
 // import ReactDOM from 'react-dom';
 // import PropTypes from 'prop-types';
 
 // const avatarChange = null;
 // avatarList= toggle chooice = [ 👧🏼👧🏻👧🏿 👧🏽🧕👸 🧑‍🦱 🧑‍🦰👦👦🏽👦🏿 👦🏼👦🏻   🦊  🐱🐶 🐵];
-
 
 // const usernameChange = null;
 // input
@@ -13,18 +12,24 @@ import '../App.css';
 // const colorChange = null;
 // colorList = [red, blue, pink, yellow, black, green, purple]
 
-
-function Settings() {  
-    return(
-            <div>
-            <nav>
-                <li><p>Avatar change</p></li>
-                <li> <p>Change Username</p></li>
-                <li> <p>Change background color</p></li>
-       
-            </nav>
-        </div>
-    )
+function Settings() {
+  return (
+    <div>
+      <nav>
+        <li>
+          <p>Avatar change</p>
+        </li>
+        <li>
+          {" "}
+          <p>Change Username</p>
+        </li>
+        <li>
+          {" "}
+          <p>Change background color</p>
+        </li>
+      </nav>
+    </div>
+  );
 }
 
 export default Settings;
