--- src/components/error-screen.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-// import ReactDOM from 'react-dom';
-// import PropTypes from 'prop-types';
-
-const errorScreen = (
-    <div>
-    <span>We're sorry. Something must have went wrong with the backend script. We apologize for the inconvenience</span>
-    </div>
-)
-
-
-export default errorScreen;
