--- src/components/followers.jsx
@@ -1,90 +1,0 @@
-import '../App.css';
-// import { useState } from "react";
-// import ReactDOM from 'react-dom';
-// import PropTypes from 'prop-types';
-
-// function block() {
-//     alert('User blocked');
-// }
-//<span> <button className="userOption" onClick={block}>✋</button></span>Block 
-
-// one article section
-
-function followers() {  
-    return(
-            <div>
-                <main className="followArticle">
-            <article className="followsStyle">
-            <span className="avatarStyle">👦🏻  </span>
-                <span>bgrgu121</span>
-                <br/>
-                <span>•Troublemaker</span>
-                <br/>
-                <span>•Livin’ a little</span>
-            </article>
-            <article className="followsStyle">
-            <span className="avatarStyle">👸  </span>
-                <span>wrbzd</span> 
-                <br/>
-                <span>•Designer</span>
-                <br/>
-                <span> •I’m real and I hope some of my followers are too.</span>
-            </article>
-            <article className="followsStyle">
-                <span className="avatarStyle">🧕  </span>
-                <span>pxtfh</span>
-                <br/>
-                <span>•Photographer</span>
-                <br/>
-                <span> •You miss 100% of the dogs you don’t pet.</span>
-            </article>
-            <article className="followsStyle">
-            <span className="avatarStyle">👦🏼   </span>
-                <span>hpyge</span>
-                <br/>
-                <span>•Total entrepreneur</span>
-                <br/>
-                <span>•Stay humble. Be kind. Work hard.</span>
-            </article>
-
-            <article className="followsStyle">
-                <span className="avatarStyle">🦊  </span>
-                <span>qqxcd31</span>
-                <br/>
-                <span>•Reader</span>
-                <br/>
-                <span>•Hot and dangerous</span>
-            </article>
-
-            <article className="followsStyle">
-            <span className="avatarStyle">👧🏿  </span>
-                <span>pfngye2422</span>
-                <br/>
-                <span>•Typical communicator</span>
-                <br/>
-                <span>•Keepin' it real since 2000</span>
-            </article>
-
-            <article className="followsStyle">
-            <span className="avatarStyle">👧🏻  </span>
-                <span>U2wp67</span>
-                <br/>
-                <span>•Award-winning thinker</span>
-                <br/>
-                <span>•Viewer discretion advised</span>
-            </article>
-
-            <article className="followsStyle">
-            <span className="avatarStyle">👦🏽  </span>
-                <span>2kyrke3</span>
-                <br/>
-                <span>•Introvert</span>
-                <br/>
-                <span>•🍩worry 🐝 happy</span>
-            </article>
-        </main>
-        </div>
-    )
-}
-
-export default followers;
