+++ src/index.js
@@ -1,12 +1,12 @@
 import React from "react";
- import ReactDOM from "react-dom";
- import "./index.css";
+import ReactDOM from "react-dom";
+import "./index.css";
 import Home from "./Home";
- import reportWebVitals from "./reportWebVitals";
- 
- ReactDOM.render(
-   <React.StrictMode>
+import reportWebVitals from "./reportWebVitals";
+
+ReactDOM.render(
+  <React.StrictMode>
     <Home />
-   </React.StrictMode>,
-   document.getElementById("root")
- );
+  </React.StrictMode>,
+  document.getElementById("root")
+);
